+++ src/App.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { AuthProvider } from './Components/ComponentsPages/AuthContext';
-import Login from './Components/ComponentsPages/Login';
-import ProtectedRoute from './Components/ProtectedRoute';
-import Profile from './Components/ComponentsPages/Profile';
+import Login from './Components/ComponentsUser/Login';
 import Home from './Components/ComponentsPages/Home';
 import LandingPage from './Components/ComponentsPages/LandingPage';
-import Register from './Components/ComponentsPages/Register';
+import Register from './Components/ComponentsUser/Register';
 
 
 const App: React.FC = () => {
@@ -18,7 +16,6 @@ const App: React.FC = () => {
         <Route path="/register" element={<Register />} />
           <Route path="/login" element={<Login />} />
           <Route path = "/home" element = {<Home/>} />
-
         </Routes>
       </Router>
     </AuthProvider>
